+++ tencentcloud/mps/v20190612/models.go
@@ -3416,6 +3416,9 @@ type DescribeTaskDetailResponse struct {
 		// 来源上下文，用于透传用户请求信息，任务流状态变更回调将返回该字段值，最长1000个字符。
 		SessionContext *string `json:"SessionContext,omitempty" name:"SessionContext"`
 
+		// 扩展信息字段，仅用于特定场景。
+		ExtInfo *string `json:"ExtInfo,omitempty" name:"ExtInfo"`
+
 		// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
 		RequestId *string `json:"RequestId,omitempty" name:"RequestId"`
 	} `json:"Response"`
@@ -4064,7 +4067,7 @@ type ImageWatermarkInputForUpdate struct {
 
 	// 水印的高度。支持 %、px 两种格式：
 	// <li>当字符串以 % 结尾，表示水印 Height 为视频高度的百分比大小，如 10% 表示 Height 为视频高度的 10%；</li>
-	// <li>当字符串以 px 结尾，表示水印 Width 单位为像素，如 100px 表示 Width 为 100 像素。取值范围为0或[8, 4096]。</li>
+	// <li>当字符串以 px 结尾，表示水印 Height 单位为像素，如 100px 表示 Height 为 100 像素。取值范围为0或[8, 4096]。</li>
 	// 默认值：0px，表示 Height 按照原始水印图片的宽高比缩放。
 	Height *string `json:"Height,omitempty" name:"Height"`
 }
@@ -4081,7 +4084,7 @@ type ImageWatermarkTemplate struct {
 
 	// 水印的高度。支持 %、px 两种格式：
 	// <li>当字符串以 % 结尾，表示水印 Height 为视频高度的百分比大小，如 10% 表示 Height 为视频高度的 10%；</li>
-	// <li>当字符串以 px 结尾，表示水印 Width 单位为像素，如 100px 表示 Width 为 100 像素；</li>
+	// <li>当字符串以 px 结尾，表示水印 Height 单位为像素，如 100px 表示 Height 为 100 像素；</li>
 	// 0px：表示 Height 按照 Width 对视频宽度的比例缩放。
 	Height *string `json:"Height,omitempty" name:"Height"`
 }
@@ -5653,7 +5656,7 @@ type ModifyTranscodeTemplateRequest struct {
 	// 转码模板名称，长度限制：64 个字符。
 	Name *string `json:"Name,omitempty" name:"Name"`
 
-	// 模板描述信息，长度限制：256 个字节。
+	// 模板描述信息，长度限制：256 个字符。
 	Comment *string `json:"Comment,omitempty" name:"Comment"`
 
 	// 是否去除视频数据，可选值：