+++ tencentcloud/vod/v20180717/models.go
@@ -5842,7 +5842,7 @@ type ImageWatermarkInput struct {
 
 	// 水印的高度。支持 %、px 两种格式：
 	// <li>当字符串以 % 结尾，表示水印 Height 为视频高度的百分比大小，如 10% 表示 Height 为视频高度的 10%；</li>
-	// <li>当字符串以 px 结尾，表示水印 Width 单位为像素，如 100px 表示 Width 为 100 像素。取值范围为0或[8, 4096]。</li>
+	// <li>当字符串以 px 结尾，表示水印 Height 单位为像素，如 100px 表示 Height 为 100 像素。取值范围为0或[8, 4096]。</li>
 	// 默认值：0px，表示 Height 按照原始水印图片的宽高比缩放。
 	Height *string `json:"Height,omitempty" name:"Height"`
 }
@@ -5859,8 +5859,7 @@ type ImageWatermarkInputForUpdate struct {
 
 	// 水印的高度。支持 %、px 两种格式：
 	// <li>当字符串以 % 结尾，表示水印 Height 为视频高度的百分比大小，如 10% 表示 Height 为视频高度的 10%；</li>
-	// <li>当字符串以 px 结尾，表示水印 Width 单位为像素，如 100px 表示 Width 为 100 像素。取值范围为0或[8, 4096]。</li>
-	// 默认值：0px，表示 Height 按照原始水印图片的宽高比缩放。
+	// <li>当字符串以 px 结尾，表示水印 Height 单位为像素，如 100px 表示 Height 为 100 像素。取值范围为0或[8, 4096]。</li>
 	Height *string `json:"Height,omitempty" name:"Height"`
 }
 
@@ -5876,7 +5875,7 @@ type ImageWatermarkTemplate struct {
 
 	// 水印的高度。支持 %、px 两种格式：
 	// <li>当字符串以 % 结尾，表示水印 Height 为视频高度的百分比大小，如 10% 表示 Height 为视频高度的 10%；</li>
-	// <li>当字符串以 px 结尾，表示水印 Width 单位为像素，如 100px 表示 Width 为 100 像素；</li>
+	// <li>当字符串以 px 结尾，表示水印 Height 单位为像素，如 100px 表示 Height 为 100 像素；</li>
 	// 0px：表示 Height 按照 Width 对视频宽度的比例缩放。
 	Height *string `json:"Height,omitempty" name:"Height"`
 }
@@ -7812,7 +7811,7 @@ type ModifyTranscodeTemplateRequest struct {
 	// 转码模板名称，长度限制：64 个字符。
 	Name *string `json:"Name,omitempty" name:"Name"`
 
-	// 模板描述信息，长度限制：256 个字节。
+	// 模板描述信息，长度限制：256 个字符。
 	Comment *string `json:"Comment,omitempty" name:"Comment"`
 
 	// 是否去除视频数据，可选值：